split main.cpp and add ambient temperature
