+++ Makefile
@@ -1,6 +1,5 @@
-
-CC     = cc
-CFLAGS = -O2 -Wall
+CC     = clang
+CFLAGS = -O3 -Wall
 INC    = -framework IOKit
 PREFIX = /usr/local
 EXEC   = osx-cpu-temp
@@ -9,18 +8,16 @@ SOURCES := $(shell find . -name '*.c')
 build : $(EXEC)
 
 clean : 
-	rm $(EXEC)
+	rm -f $(EXEC) *.o *.out
 
-$(EXEC) : smc.c
+$(EXEC) : main.cpp smc.o
 	$(CC) $(CFLAGS) $(INC) -o $@ $?
+
+%.o: %.c %.h
+	$(CC) $(CFLAGS) -c $< -o $@
 
 install : $(EXEC)
 	@install -v $(EXEC) $(PREFIX)/bin/$(EXEC)
 
-lint: bin/lint.sh $(SOURCES)
-	@$<
 
-lint-fix: bin/lint-fix.sh $(SOURCES)
-	@$<
-
-.PHONY: lint lint-fix clean install build
+.PHONY: clean install build
